fix stats calculation
The hardcoded interval was not accounting for the API latency which
seems to be about 200ms.  It's possible that the stats are still
generated when I start the function call and it just takes 200ms for the
response to be returned, but I'd rather include the 200ms and under
report the bandwidth than over report it.

The more long term fix is to just log the raw byte values and then
calculate the deltas separately.