Experimental 'two tuple host' feature
Add a new configuration option for 'two tuple hosts'

If an ACL is attempted to be added for a matching host, add it as a 2
tuple acl instead of a 4 tuple acl.

While I think the feature is needed, it should possibly be implemented
in the client or some place else and this will most likely change.